+++ backend/infrastructure/src/main/kotlin/me/pinfort/tsvideosmanager/infrastructure/database/dto/CreatedFileDto.kt
@@ -1,8 +1,8 @@
 package me.pinfort.tsvideosmanager.infrastructure.database.dto
 
 data class CreatedFileDto(
-    val id: Int,
-    val splittedFileId: Int,
+    val id: Long,
+    val splittedFileId: Long,
     val file: String,
     val size: Long,
     val mime: String?,
